requirements.txt
Flask==2.2.5
GitPython==3.1.43
psutil==5.9.0
pyjson5==1.6.6